Made Availability plugin use dist metadata info
Availability weaver plug was using a hardwired github base
URL.  Added more flexibility so that any repository URL and
repository type (defaults to git) will be picked up from the
dist metadata.

If the home page URL is not the CPAN default then an additional
paragraph with this information is added in.

If the repository is on github some additional magic is added,
otherwise a minimal generic paragraph is added.  This could later
be extended for other repositories.